+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Next
 - Supports Tauri 2
 - Supports Svelte 3, 4 and 5
+- `files` are now only provided when the user hovers over the dropzone.
 
 ## 1.0.0 - 2021 Dec 1
 - Initial release